requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.48
psycopg2-binary>=2.9.11
httpx>=0.27.0
paddleocr==2.7.3
paddlepaddle==2.6.2
opencv-python>=4.6.0
numpy==1.26.4
python-dotenv>=1.0.0
cloudinary>=1.41.0
Pillow>=10.0.0